Minor fix to used and defined registers and to destination operand logic: - Updated types to uint64_t from uint32_t to account for more registers - StackPush operations don't have a destination operand (aside from the implied stack pointer), so this commit makes this fix

+++ include/X86Instruction.h
@@ -585,8 +585,8 @@ public:
 
     bool usesFlag(uint32_t flg);
     bool defsFlag(uint32_t flg);
-    bool implicitlyUsesReg(uint32_t alu);
-    bool implicitlyDefinesReg(uint32_t alu);
+    bool implicitlyUsesReg(uint64_t alu);
+    bool implicitlyDefinesReg(uint64_t alu);
 
     struct VectorInfo getVectorInfo();
     struct RuntimeValue getRegisterValue(enum ud_type reg);
